+++ flashdreams/flashdreams/api_v2/slangpy_ui_thread.py
@@ -1,0 +1,115 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""SlangPy UI thread for FlashDreams applications."""
+
+from abc import ABC, abstractmethod
+from typing import Any, Generic, TypeVar, final
+
+from torch import Tensor
+
+from flashdreams.api_v2.thread import UIThread
+from flashdreams.runtime_v2._slangpy_ui_renderer import (
+    _SlangPyUIRenderer,
+    _UIRenderer,
+)
+from flashdreams.runtime_v2.presentation_manager import PresentationManager
+from flashdreams.runtime_v2.step_result import StepResult
+from flashdreams.runtime_v2.user_input_events import UserInputEvents
+from flashdreams.runtime_v2.video_tensor import VideoTensorLayout
+
+_StateT = TypeVar("_StateT")
+
+
+class SlangPyUIThread(UIThread[_StateT], ABC, Generic[_StateT]):
+    """Render a SlangPy UI over an optional model frame."""
+
+    def __init__(
+        self,
+        *,
+        state: _StateT,
+        frequency: int,
+        output_layout: VideoTensorLayout,
+        presentation_manager: PresentationManager,
+        renderer: _UIRenderer | None = None,
+        width: int | None = None,
+        height: int | None = None,
+    ) -> None:
+        """Configure a SlangPy UI thread without creating GPU resources.
+
+        Args:
+            state: State used by the UI thread.
+            frequency: Maximum UI iterations per second.
+            output_layout: Layout used for the compositing result.
+            presentation_manager: Buffer containing model frames.
+            renderer: Rendering backend; ``None`` creates the SlangPy backend.
+            width: Render-target width, required for the default renderer.
+            height: Render-target height, required for the default renderer.
+
+        Raises:
+            ValueError: ``output_layout`` is not ``tchw``, or the default
+                renderer has no output dimensions.
+        """
+        if output_layout is not VideoTensorLayout.tchw:
+            raise ValueError("SlangPy UI rendering requires tchw output.")
+        super().__init__(
+            state=state,
+            frequency=frequency,
+            output_layout=output_layout,
+            presentation_manager=presentation_manager,
+        )
+        if renderer is None:
+            if width is None or height is None:
+                raise ValueError(
+                    "width and height are required when renderer is not supplied."
+                )
+            renderer = _SlangPyUIRenderer(width=width, height=height)
+        self.renderer = renderer
+
+    @abstractmethod
+    def draw_ui(
+        self, ui: Any, step_index: int, events: UserInputEvents
+    ) -> Tensor | None:
+        """Draw widgets and optionally return the frame beneath them."""
+        ...
+
+    @final
+    def step_ui(self, step_index: int, events: UserInputEvents) -> StepResult:
+        """Render the UI over the optional back-buffer returned by :meth:`draw_ui`."""
+        back_buffer: Tensor | None = None
+
+        def draw(ui: Any, index: int, current_events: UserInputEvents) -> None:
+            nonlocal back_buffer
+            back_buffer = self.draw_ui(ui, index, current_events)
+
+        overlay = self.renderer.render(step_index, events, draw)
+        frame = self._presentation_manager.composite(back_buffer, overlay)
+        return StepResult(
+            step_index=step_index,
+            output=frame.unsqueeze(0),
+            frame_count=1,
+            output_layout=self.output_layout,
+        )
+
+    def reset(self) -> None:
+        """Reset renderer state after a session reset event."""
+        self.renderer.reset()
+
+    def close(self) -> None:
+        """Release the renderer."""
+        self.renderer.close()
+
+
+__all__ = ["SlangPyUIThread"]
