+++ integrations_v2/slangpy_ui_demo/pyproject.toml
@@ -6,9 +6,9 @@ requires = ["setuptools>=69", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "flashdreams-imgui-demo"
+name = "flashdreams-slangpy-ui-demo"
 version = "0.1.0"
-description = "Dear ImGui demos for the FlashDreams v2 threaded runtime."
+description = "SlangPy UI demos for the FlashDreams v2 threaded runtime."
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
@@ -16,14 +16,14 @@ dependencies = [
 ]
 
 [project.entry-points."flashdreams.applications_v2"]
-"imgui-text-input" = "imgui_demo.text_input_app:create_app"
-"imgui-model-output" = "imgui_demo.model_output_app:create_app"
+"slangpy-ui-text-input" = "slangpy_ui_demo.text_input_app:create_app"
+"slangpy-ui-model-output" = "slangpy_ui_demo.model_output_app:create_app"
 
 [tool.uv.sources]
 flashdreams = { workspace = true }
 
 [tool.setuptools.packages.find]
-include = ["imgui_demo*"]
+include = ["slangpy_ui_demo*"]
 
 [tool.uv]
 managed = true
