+++ integrations_v2/slangpy_ui_demo/slangpy_ui_demo/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Interactive Dear ImGui applications for the v2 threaded runtime."""
+"""Interactive SlangPy UI applications for the v2 threaded runtime."""
 
 from .model_output_app import ModelOutputApplication, ModelOutputSession
 from .text_input_app import TextInputApplication, TextInputSession, create_app
